--- src/main/java/com/pragma/powerup/usermicroservice/adapters/driven/jpa/mysql/entity/PersonEntity.java
@@ -1,35 +1,0 @@
-package com.pragma.powerup.usermicroservice.adapters.driven.jpa.mysql.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "person")
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class PersonEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    private String surname;
-    private String mail;
-    private String phone;
-    private String address;
-    private String idDniType;
-    @Column(unique = true, nullable = false, length = 20)
-    private String dniNumber;
-    private String idPersonType;
-    private String password;
-    private String tokenPassword;
-}
