+++ src/main/java/com/pragma/powerup/usermicroservice/adapters/driven/jpa/mysql/entity/PrincipalUser.java
@@ -24,7 +24,7 @@ public class PrincipalUser implements UserDetails {
         this.authorities = authorities;
     }
 
-    public static PrincipalUser build(PersonEntity usuario, List<RoleEntity> roles) {
+    public static PrincipalUser build(UserEntity usuario, List<RoleEntity> roles) {
         List<GrantedAuthority> authorities = roles.stream()
                 .map(rol -> new SimpleGrantedAuthority(rol.getName())).collect(Collectors.toList());
         return new PrincipalUser(usuario.getName(), usuario.getDniNumber(), usuario.getMail(),
