+++ src/main/java/com/pragma/powerup/usermicroservice/adapters/driven/jpa/mysql/entity/UserEntity.java
@@ -1,5 +1,6 @@
 package com.pragma.powerup.usermicroservice.adapters.driven.jpa.mysql.entity;
 
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -22,10 +23,18 @@ public class UserEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @ManyToOne
-    @JoinColumn(name = "id_person")
-    private PersonEntity personEntity;
-    @ManyToOne
+    private String name;
+    private String surname;
+    private String mail;
+    private String phone;
+    private String address;
+    private String idDniType;
+    @Column(unique = true, nullable = false, length = 20)
+    private String dniNumber;
+    private String idPersonType;
+    private String password;
+    private String tokenPassword;
+    @ManyToOne(optional = true)
     @JoinColumn(name = "id_role")
-    private RoleEntity roleEntity;
+    private RoleEntity role;
 }
