--- src/main/java/com/pragma/powerup/usermicroservice/adapters/driven/jpa/mysql/mappers/IPersonEntityMapper.java
@@ -1,13 +1,0 @@
-package com.pragma.powerup.usermicroservice.adapters.driven.jpa.mysql.mappers;
-
-import com.pragma.powerup.usermicroservice.adapters.driven.jpa.mysql.entity.PersonEntity;
-import com.pragma.powerup.usermicroservice.domain.model.Person;
-import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
-
-@Mapper(componentModel = "spring",
-        unmappedTargetPolicy = ReportingPolicy.IGNORE,
-        unmappedSourcePolicy = ReportingPolicy.IGNORE)
-public interface IPersonEntityMapper {
-    PersonEntity toEntity(Person person);
-}
