--- src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/controller/PersonRestController.java
@@ -1,39 +1,0 @@
-package com.pragma.powerup.usermicroservice.adapters.driving.http.controller;
-
-import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.PersonRequestDto;
-import com.pragma.powerup.usermicroservice.adapters.driving.http.handlers.IPersonHandler;
-import com.pragma.powerup.usermicroservice.configuration.Constants;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Collections;
-import java.util.Map;
-
-@RestController
-@RequestMapping("/person/")
-@RequiredArgsConstructor
-public class PersonRestController {
-    private final IPersonHandler personHandler;
-
-    @Operation(summary = "Add a new person",
-            responses = {
-                @ApiResponse(responseCode = "201", description = "Person created",
-                        content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Map"))),
-                @ApiResponse(responseCode = "409", description = "Person already exists",
-                        content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Error")))})
-    @PostMapping
-    public ResponseEntity<Map<String, String>> savePerson(@RequestBody PersonRequestDto personRequestDto) {
-        personHandler.savePerson(personRequestDto);
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .body(Collections.singletonMap(Constants.RESPONSE_MESSAGE_KEY, Constants.PERSON_CREATED_MESSAGE));
-    }
-}
