+++ src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/controller/UserRestController.java
@@ -1,103 +1,39 @@
 package com.pragma.powerup.usermicroservice.adapters.driving.http.controller;
 
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.UserRequestDto;
-import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.PersonResponseDto;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.handlers.IUserHandler;
 import com.pragma.powerup.usermicroservice.configuration.Constants;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 
 @RestController
-@RequestMapping("/user")
+@RequestMapping("/user/")
 @RequiredArgsConstructor
-@SecurityRequirement(name = "jwt")
 public class UserRestController {
-    private final IUserHandler userHandler;
+    private final IUserHandler personHandler;
 
     @Operation(summary = "Add a new user",
             responses = {
-                    @ApiResponse(responseCode = "201", description = "User created",
-                            content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Map"))),
-                    @ApiResponse(responseCode = "409", description = "User already exists",
-                            content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Error"))),
-                    @ApiResponse(responseCode = "403", description = "Role not allowed for user creation",
-                            content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Error")))})
-    @PostMapping("")
-    public ResponseEntity<Map<String, String>> saveUser(@RequestBody UserRequestDto userRequestDto) {
-        userHandler.saveUser(userRequestDto);
+                @ApiResponse(responseCode = "201", description = "Person created",
+                        content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Map"))),
+                @ApiResponse(responseCode = "409", description = "Person already exists",
+                        content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Error")))})
+    @PostMapping
+    public ResponseEntity<Map<String, String>> savePerson(@RequestBody UserRequestDto userRequestDto) {
+        personHandler.saveUser(userRequestDto);
         return ResponseEntity.status(HttpStatus.CREATED)
-                .body(Collections.singletonMap(Constants.RESPONSE_MESSAGE_KEY, Constants.USER_CREATED_MESSAGE));
-    }
-    @Operation(summary = "Delete an user",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "User deleted",
-                            content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Map"))),
-                    @ApiResponse(responseCode = "404", description = "User not found",
-                            content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Error")))})
-    @DeleteMapping("")
-    public ResponseEntity<Map<String, String>> deleteUser(@RequestBody UserRequestDto userRequestDto) {
-        userHandler.deleteUser(userRequestDto);
-        return ResponseEntity.ok(Collections.singletonMap(Constants.RESPONSE_MESSAGE_KEY, Constants.USER_DELETED_MESSAGE));
-    }
-    @Operation(summary = "Get all the providers",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "All providers returned",
-                            content = @Content(mediaType = "application/json",
-                                    array = @ArraySchema(schema = @Schema(implementation = PersonResponseDto.class)))),
-                    @ApiResponse(responseCode = "404", description = "No data found",
-                            content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Error")))})
-    @GetMapping("/provider")
-    public ResponseEntity<List<PersonResponseDto>> getAllProviders(@Parameter(description = "Number of the page to list providers") @RequestParam int page) {
-        return ResponseEntity.ok(userHandler.getProvider(page));
-    }
-    @Operation(summary = "Get a provider user",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "Provider user returned",
-                            content = @Content(mediaType = "application/json", schema = @Schema(implementation = PersonResponseDto.class))),
-                    @ApiResponse(responseCode = "404", description = "User not found with provider role",
-                            content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Error")))})
-    @GetMapping("/provider/{id}")
-    public ResponseEntity<PersonResponseDto> getProvider(@PathVariable Long id) {
-        return ResponseEntity.ok(userHandler.getProvider(id));
-    }
-    @Operation(summary = "Get a employee user",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "Employee user returned",
-                            content = @Content(mediaType = "application/json", schema = @Schema(implementation = PersonResponseDto.class))),
-                    @ApiResponse(responseCode = "404", description = "User not found with employee role",
-                            content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Error")))})
-    @GetMapping("/employee/{id}")
-    public ResponseEntity<PersonResponseDto> getEmployee(@PathVariable Long id) {
-        return ResponseEntity.ok(userHandler.getEmployee(id));
-    }
-    @Operation(summary = "Get a client user",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "Client user returned",
-                            content = @Content(mediaType = "application/json", schema = @Schema(implementation = PersonResponseDto.class))),
-                    @ApiResponse(responseCode = "404", description = "User not found with client role",
-                            content = @Content(mediaType = "application/json", schema = @Schema(ref = "#/components/schemas/Error")))})
-    @GetMapping("/client/{id}")
-    public ResponseEntity<PersonResponseDto> getClient(@PathVariable Long id) {
-        return ResponseEntity.ok(userHandler.getClient(id));
+                .body(Collections.singletonMap(Constants.RESPONSE_MESSAGE_KEY, Constants.PERSON_CREATED_MESSAGE));
     }
 }
