--- src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/dto/request/PersonRequestDto.java
@@ -1,18 +1,0 @@
-package com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class PersonRequestDto {
-    private String name;
-    private String surname;
-    private String mail;
-    private String phone;
-    private String address;
-    private String idDniType;
-    private String dniNumber;
-    private String idPersonType;
-    private String password;
-}
