--- src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/dto/response/PersonResponseDto.java
@@ -1,17 +1,0 @@
-package com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class PersonResponseDto {
-    private String name;
-    private String surname;
-    private String mail;
-    private String phone;
-    private String address;
-    private String idDniType;
-    private String dniNumber;
-    private String idPersonType;
-}
