--- src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/handlers/IPersonHandler.java
@@ -1,7 +1,0 @@
-package com.pragma.powerup.usermicroservice.adapters.driving.http.handlers;
-
-import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.PersonRequestDto;
-
-public interface IPersonHandler {
-    void savePerson(PersonRequestDto personRequestDto);
-}
