--- src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/mapper/IPersonRequestMapper.java
@@ -1,13 +1,0 @@
-package com.pragma.powerup.usermicroservice.adapters.driving.http.mapper;
-
-import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.PersonRequestDto;
-import com.pragma.powerup.usermicroservice.domain.model.Person;
-import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
-
-@Mapper(componentModel = "spring",
-        unmappedTargetPolicy = ReportingPolicy.IGNORE,
-        unmappedSourcePolicy = ReportingPolicy.IGNORE)
-public interface IPersonRequestMapper {
-    Person toPerson(PersonRequestDto personRequestDto);
-}
