+++ src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/mapper/IUserRequestMapper.java
@@ -3,14 +3,11 @@ package com.pragma.powerup.usermicroservice.adapters.driving.http.mapper;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.UserRequestDto;
 import com.pragma.powerup.usermicroservice.domain.model.User;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 import org.mapstruct.ReportingPolicy;
 
 @Mapper(componentModel = "spring",
         unmappedTargetPolicy = ReportingPolicy.IGNORE,
         unmappedSourcePolicy = ReportingPolicy.IGNORE)
 public interface IUserRequestMapper {
-    @Mapping(target = "person.id", source = "idPerson")
-    @Mapping(target = "role.id", source = "idRole")
-    User toUser(UserRequestDto userRequestDto);
+    User toUser(UserRequestDto personRequestDto);
 }
