+++ src/main/java/com/pragma/powerup/usermicroservice/configuration/security/MainSecurity.java
@@ -46,8 +46,8 @@ public class MainSecurity {
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
         http.cors().and().csrf().disable()
                 .authorizeRequests(requests -> requests
-                        .requestMatchers("/auth/login", "/swagger-ui.html", "/swagger-ui/**", "/v3/api-docs/**", "/actuator/health", "/person/").permitAll()
-                        .requestMatchers("/user").hasRole("ADMIN")
+                        .requestMatchers("/auth/login", "/swagger-ui.html", "/swagger-ui/**", "/v3/api-docs/**", "/actuator/health", "/user/").permitAll()
+                        //.requestMatchers("/user").hasRole("ADMIN")
                         .anyRequest().authenticated()
                 )
                 .formLogin().disable()
