--- src/main/java/com/pragma/powerup/usermicroservice/domain/api/IPersonServicePort.java
@@ -1,7 +1,0 @@
-package com.pragma.powerup.usermicroservice.domain.api;
-
-import com.pragma.powerup.usermicroservice.domain.model.Person;
-
-public interface IPersonServicePort {
-    void savePerson(Person person);
-}
