+++ src/main/java/com/pragma/powerup/usermicroservice/domain/api/IUserServicePort.java
@@ -2,13 +2,6 @@ package com.pragma.powerup.usermicroservice.domain.api;
 
 import com.pragma.powerup.usermicroservice.domain.model.User;
 
-import java.util.List;
-
 public interface IUserServicePort {
     void saveUser(User user);
-    void deleteUser(User user);
-    List<User> getAllProviders(int page);
-    User getProvider(Long id);
-    User getEmployee(Long id);
-    User getClient(Long id);
 }
