+++ src/main/java/com/pragma/powerup/usermicroservice/domain/model/User.java
@@ -2,13 +2,27 @@ package com.pragma.powerup.usermicroservice.domain.model;
 
 public class User {
     private Long id;
-    private Person person;
-    private Role role;
+    private String name;
+    private String surname;
+    private String mail;
+    private String phone;
+    private String address;
+    private String idDniType;
+    private String dniNumber;
+    private String idPersonType;
+    private String password;
 
-    public User(Long id, Person person, Role role) {
+    public User(Long id, String name, String surname, String mail, String phone, String address, String idDniType, String dniNumber, String idPersonType, String password) {
         this.id = id;
-        this.person = person;
-        this.role = role;
+        this.name = name;
+        this.surname = surname;
+        this.mail = mail;
+        this.phone = phone;
+        this.address = address;
+        this.idDniType = idDniType;
+        this.dniNumber = dniNumber;
+        this.idPersonType = idPersonType;
+        this.password = password;
     }
 
     public Long getId() {
@@ -19,19 +33,75 @@ public class User {
         this.id = id;
     }
 
-    public Person getPerson() {
-        return person;
+    public String getName() {
+        return name;
     }
 
-    public void setPerson(Person person) {
-        this.person = person;
+    public void setName(String name) {
+        this.name = name;
     }
 
-    public Role getRole() {
-        return role;
+    public String getSurname() {
+        return surname;
     }
 
-    public void setRole(Role role) {
-        this.role = role;
+    public void setSurname(String surname) {
+        this.surname = surname;
+    }
+
+    public String getMail() {
+        return mail;
+    }
+
+    public void setMail(String mail) {
+        this.mail = mail;
+    }
+
+    public String getPhone() {
+        return phone;
+    }
+
+    public void setPhone(String phone) {
+        this.phone = phone;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public String getIdDniType() {
+        return idDniType;
+    }
+
+    public void setIdDniType(String idDniType) {
+        this.idDniType = idDniType;
+    }
+
+    public String getDniNumber() {
+        return dniNumber;
+    }
+
+    public void setDniNumber(String dniNumber) {
+        this.dniNumber = dniNumber;
+    }
+
+    public String getIdPersonType() {
+        return idPersonType;
+    }
+
+    public void setIdPersonType(String idPersonType) {
+        this.idPersonType = idPersonType;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
     }
 }
