--- src/main/java/com/pragma/powerup/usermicroservice/domain/spi/IPersonPersistencePort.java
@@ -1,7 +1,0 @@
-package com.pragma.powerup.usermicroservice.domain.spi;
-
-import com.pragma.powerup.usermicroservice.domain.model.Person;
-
-public interface IPersonPersistencePort {
-    void savePerson(Person person);
-}
