+++ src/main/resources/application-dev.yml
@@ -4,8 +4,8 @@ server:
 spring:
   datasource:
     url: jdbc:mysql://localhost/powerup
-    username: root
-    password: 1234
+    username: powerup
+    password: powerup
   jpa:
     hibernate:
       ddl-auto: update
